// src/enums/store/movie.enum.ts
export enum EGenre {
  Action = "Action",
  Thriller = "Thriller",
  Romance = "Romance",
  Drama = "Drama",
  Crime = "Crime",
  Biography = "Biography",
  SciFi = "SciFi",
}

export enum EType {
  TVseries = "TVseries",
  Movie = "Movie",
}
export enum EMovieGetters {
  getAllMovies = "getAllMovies",
  getMoviesByGenre = "getMoviesByGenre",
  getMoviesByType = "getMoviesByType",
  getFeaturedMovies = "getFeaturedMovies",
}